add vmtest config with kernel that has bridge support, simplify jiffy conversion
